Consolidate #[derive(FromVariant, ToVariant)] tests into single file
